3번째 뷰까지 완료 및 사진 넣음

+++ app/src/main/java/com/example/imagecup/ui/MainViewModel.kt
@@ -166,7 +166,7 @@ class MainViewModel @Inject constructor(
         }
     }
 
-    fun ratePhoto(getPhotosResponse: GetPhotosResponse, score: Int) {
+    fun ratePhoto(getPhotosResponse: GetPhotosResponse, score: Float) {
         viewModelScope.launch {
             repository.evaluationPhoto(getPhotosResponse.photoId, PrefsManager.uid, score)
                 .collectLatest {
